Remove legacy contact searching
