+++ temba/contacts/omnibox.py
@@ -7,9 +7,10 @@ from django.db.models.functions import Upper
 
 from temba.channels.models import Channel
 from temba.contacts.models import Contact, ContactGroup, ContactGroupCount, ContactURN
-from temba.contacts.search import SearchException, search_contacts
 from temba.msgs.models import Label
 from temba.utils.models import IDSliceQuerySet
+
+from .search import SearchException, search_contacts
 
 SEARCH_ALL_GROUPS = "g"
 SEARCH_STATIC_GROUPS = "s"
@@ -93,7 +94,7 @@ def omnibox_mixed_search(org, query, types):
 
     if SEARCH_CONTACTS in search_types:
         try:
-            search_results = search_contacts(org.id, org.cached_active_contacts_group.uuid, query, "name")
+            search_results = search_contacts(org, query, group=org.cached_active_contacts_group, sort="name")
             contacts = IDSliceQuerySet(Contact, search_results.contact_ids, 0, len(search_results.contact_ids))
             results += list(contacts[:per_type_limit])
             Contact.bulk_cache_initialize(org, contacts=results)
@@ -104,10 +105,12 @@ def omnibox_mixed_search(org, query, types):
     if SEARCH_URNS in search_types:
         if not org.is_anon and query and len(query) >= 3:
             try:
-                # build an ORed query of all sendable schemes
+                # build an OR'ed query of all sendable schemes
                 sendable_schemes = org.get_schemes(Channel.ROLE_SEND)
                 scheme_query = " OR ".join(f"{s} ~ {json.dumps(query)}" for s in sendable_schemes)
-                search_results = search_contacts(org.id, org.cached_active_contacts_group.uuid, scheme_query, "name")
+                search_results = search_contacts(
+                    org, scheme_query, group=org.cached_active_contacts_group, sort="name"
+                )
                 urns = ContactURN.objects.filter(
                     contact_id__in=search_results.contact_ids, scheme__in=sendable_schemes
                 )
