+++ temba/contacts/search/__init__.py
@@ -1,1 +1,1 @@
-from .parser import *  # noqa
+from .base import ParsedQuery, SearchException, SearchResults, parse_query, search_contacts  # noqa
