--- temba/contacts/search/gen/ContactQL.interp
@@ -1,32 +1,0 @@
-token literal names:
-null
-'('
-')'
-null
-null
-null
-null
-null
-null
-null
-
-token symbolic names:
-null
-LPAREN
-RPAREN
-AND
-OR
-COMPARATOR
-TEXT
-STRING
-WS
-ERROR
-
-rule names:
-parse
-expression
-literal
-
-
-atn:
-[3, 24715, 42794, 33075, 47597, 16764, 15335, 30598, 22884, 3, 11, 40, 4, 2, 9, 2, 4, 3, 9, 3, 4, 4, 9, 4, 3, 2, 3, 2, 3, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 5, 3, 21, 10, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 7, 3, 31, 10, 3, 12, 3, 14, 3, 34, 11, 3, 3, 4, 3, 4, 5, 4, 38, 10, 4, 3, 4, 2, 3, 4, 5, 2, 4, 6, 2, 2, 2, 42, 2, 8, 3, 2, 2, 2, 4, 20, 3, 2, 2, 2, 6, 37, 3, 2, 2, 2, 8, 9, 5, 4, 3, 2, 9, 10, 7, 2, 2, 3, 10, 3, 3, 2, 2, 2, 11, 12, 8, 3, 1, 2, 12, 13, 7, 3, 2, 2, 13, 14, 5, 4, 3, 2, 14, 15, 7, 4, 2, 2, 15, 21, 3, 2, 2, 2, 16, 17, 7, 8, 2, 2, 17, 18, 7, 7, 2, 2, 18, 21, 5, 6, 4, 2, 19, 21, 7, 8, 2, 2, 20, 11, 3, 2, 2, 2, 20, 16, 3, 2, 2, 2, 20, 19, 3, 2, 2, 2, 21, 32, 3, 2, 2, 2, 22, 23, 12, 8, 2, 2, 23, 24, 7, 5, 2, 2, 24, 31, 5, 4, 3, 9, 25, 26, 12, 7, 2, 2, 26, 31, 5, 4, 3, 8, 27, 28, 12, 6, 2, 2, 28, 29, 7, 6, 2, 2, 29, 31, 5, 4, 3, 7, 30, 22, 3, 2, 2, 2, 30, 25, 3, 2, 2, 2, 30, 27, 3, 2, 2, 2, 31, 34, 3, 2, 2, 2, 32, 30, 3, 2, 2, 2, 32, 33, 3, 2, 2, 2, 33, 5, 3, 2, 2, 2, 34, 32, 3, 2, 2, 2, 35, 38, 7, 8, 2, 2, 36, 38, 7, 9, 2, 2, 37, 35, 3, 2, 2, 2, 37, 36, 3, 2, 2, 2, 38, 7, 3, 2, 2, 2, 6, 20, 30, 32, 37]
