--- temba/contacts/search/gen/ContactQLLexer.py
@@ -1,211 +1,0 @@
-# Generated from ContactQL.g4 by ANTLR 4.7.2
-import sys
-from io import StringIO
-from typing.io import TextIO
-
-from antlr4 import *
-
-
-def serializedATN():
-    with StringIO() as buf:
-        buf.write("\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\2\13")
-        buf.write("t\b\1\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7")
-        buf.write("\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r\4\16")
-        buf.write("\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22\t\22\4\23\t\23")
-        buf.write("\3\2\3\2\3\2\3\2\3\3\3\3\3\3\3\4\3\4\3\5\3\5\3\6\3\6\3")
-        buf.write("\6\3\6\3\7\3\7\3\7\3\b\3\b\3\b\3\b\3\b\3\b\3\b\3\b\3\b")
-        buf.write("\3\b\3\b\5\bE\n\b\3\t\3\t\3\t\6\tJ\n\t\r\t\16\tK\3\n\3")
-        buf.write("\n\3\n\3\n\7\nR\n\n\f\n\16\nU\13\n\3\n\3\n\3\13\6\13Z")
-        buf.write("\n\13\r\13\16\13[\3\13\3\13\3\f\3\f\3\r\3\r\3\r\3\r\3")
-        buf.write("\r\5\rg\n\r\3\16\3\16\3\17\3\17\3\20\3\20\3\21\3\21\3")
-        buf.write("\22\3\22\3\23\3\23\2\2\24\3\2\5\2\7\3\t\4\13\5\r\6\17")
-        buf.write("\7\21\b\23\t\25\n\27\13\31\2\33\2\35\2\37\2!\2#\2%\2\3")
-        buf.write("\2\24\4\2JJjj\4\2CCcc\4\2UUuu\4\2KKkk\4\2PPpp\4\2FFff")
-        buf.write("\4\2QQqq\4\2TTtt\4\2>>@@\6\2--/\61BBaa\3\2$$\5\2\13\f")
-        buf.write('\17\17""T\2C\\\u00c2\u00d8\u00da\u00e0\u0102\u0138\u013b')
-        buf.write("\u0149\u014c\u017f\u0183\u0184\u0186\u018d\u0190\u0193")
-        buf.write("\u0195\u0196\u0198\u019a\u019e\u019f\u01a1\u01a2\u01a4")
-        buf.write("\u01ab\u01ae\u01b5\u01b7\u01be\u01c6\u01cf\u01d1\u01dd")
-        buf.write("\u01e0\u01f0\u01f3\u01f6\u01f8\u01fa\u01fc\u0234\u023c")
-        buf.write("\u023d\u023f\u0240\u0243\u0248\u024a\u0250\u0372\u0374")
-        buf.write("\u0378\u0381\u0388\u038c\u038e\u03a3\u03a5\u03ad\u03d1")
-        buf.write("\u03d6\u03da\u03f0\u03f6\u03f9\u03fb\u03fc\u03ff\u0431")
-        buf.write("\u0462\u0482\u048c\u04cf\u04d2\u0530\u0533\u0558\u10a2")
-        buf.write("\u10c7\u10c9\u10cf\u1e02\u1e96\u1ea0\u1f00\u1f0a\u1f11")
-        buf.write("\u1f1a\u1f1f\u1f2a\u1f31\u1f3a\u1f41\u1f4a\u1f4f\u1f5b")
-        buf.write("\u1f61\u1f6a\u1f71\u1fba\u1fbd\u1fca\u1fcd\u1fda\u1fdd")
-        buf.write("\u1fea\u1fee\u1ffa\u1ffd\u2104\u2109\u210d\u210f\u2112")
-        buf.write("\u2114\u2117\u211f\u2126\u212f\u2132\u2135\u2140\u2141")
-        buf.write("\u2147\u2185\u2c02\u2c30\u2c62\u2c66\u2c69\u2c72\u2c74")
-        buf.write("\u2c77\u2c80\u2c82\u2c84\u2ce4\u2ced\u2cef\u2cf4\ua642")
-        buf.write("\ua644\ua66e\ua682\ua69c\ua724\ua730\ua734\ua770\ua77b")
-        buf.write("\ua788\ua78d\ua78f\ua792\ua794\ua798\ua7af\ua7b2\ua7b3")
-        buf.write("\uff23\uff3cS\2c|\u00b7\u00f8\u00fa\u0101\u0103\u0179")
-        buf.write("\u017c\u0182\u0185\u0187\u018a\u0194\u0197\u019d\u01a0")
-        buf.write("\u01a3\u01a5\u01a7\u01aa\u01af\u01b2\u01b6\u01b8\u01c1")
-        buf.write("\u01c8\u01ce\u01d0\u01f5\u01f7\u01fb\u01fd\u023b\u023e")
-        buf.write("\u0244\u0249\u0295\u0297\u02b1\u0373\u0375\u0379\u037f")
-        buf.write("\u0392\u03d0\u03d2\u03d3\u03d7\u03d9\u03db\u03f5\u03f7")
-        buf.write("\u0461\u0463\u0483\u048d\u04c1\u04c4\u0531\u0563\u0589")
-        buf.write("\u1d02\u1d2d\u1d6d\u1d79\u1d7b\u1d9c\u1e03\u1e9f\u1ea1")
-        buf.write("\u1f09\u1f12\u1f17\u1f22\u1f29\u1f32\u1f39\u1f42\u1f47")
-        buf.write("\u1f52\u1f59\u1f62\u1f69\u1f72\u1f7f\u1f82\u1f89\u1f92")
-        buf.write("\u1f99\u1fa2\u1fa9\u1fb2\u1fb6\u1fb8\u1fb9\u1fc0\u1fc6")
-        buf.write("\u1fc8\u1fc9\u1fd2\u1fd5\u1fd8\u1fd9\u1fe2\u1fe9\u1ff4")
-        buf.write("\u1ff6\u1ff8\u1ff9\u210c\u2115\u2131\u213b\u213e\u213f")
-        buf.write("\u2148\u214b\u2150\u2186\u2c32\u2c60\u2c63\u2c6e\u2c73")
-        buf.write("\u2c7d\u2c83\u2cee\u2cf0\u2cf5\u2d02\u2d27\u2d29\u2d2f")
-        buf.write("\ua643\ua66f\ua683\ua69d\ua725\ua733\ua735\ua77a\ua77c")
-        buf.write("\ua77e\ua781\ua789\ua78e\ua790\ua793\ua797\ua799\ua7ab")
-        buf.write("\ua7fc\uab5c\uab66\uab67\ufb02\ufb08\ufb15\ufb19\uff43")
-        buf.write("\uff5c\b\2\u01c7\u01cd\u01f4\u1f91\u1f9a\u1fa1\u1faa\u1fb1")
-        buf.write("\u1fbe\u1fce\u1ffe\u1ffe#\2\u02b2\u02c3\u02c8\u02d3\u02e2")
-        buf.write("\u02e6\u02ee\u02f0\u0376\u037c\u055b\u0642\u06e7\u06e8")
-        buf.write("\u07f6\u07f7\u07fc\u081c\u0826\u082a\u0973\u0e48\u0ec8")
-        buf.write("\u10fe\u17d9\u1845\u1aa9\u1c7f\u1d2e\u1d6c\u1d7a\u1dc1")
-        buf.write("\u2073\u2081\u2092\u209e\u2c7e\u2c7f\u2d71\u2e31\u3007")
-        buf.write("\u3037\u303d\u3100\ua017\ua4ff\ua60e\ua681\ua69e\ua69f")
-        buf.write("\ua719\ua721\ua772\ua78a\ua7fa\ua7fb\ua9d1\ua9e8\uaa72")
-        buf.write("\uaadf\uaaf5\uaaf6\uab5e\uab61\uff72\uffa1\u00ec\2\u00ac")
-        buf.write("\u00bc\u01bd\u01c5\u0296\u05ec\u05f2\u05f4\u0622\u0641")
-        buf.write("\u0643\u064c\u0670\u0671\u0673\u06d5\u06d7\u06fe\u0701")
-        buf.write("\u0712\u0714\u0731\u074f\u07a7\u07b3\u07ec\u0802\u0817")
-        buf.write("\u0842\u085a\u08a2\u08b4\u0906\u093b\u093f\u0952\u095a")
-        buf.write("\u0963\u0974\u0982\u0987\u098e\u0991\u0992\u0995\u09aa")
-        buf.write("\u09ac\u09b2\u09b4\u09bb\u09bf\u09d0\u09de\u09df\u09e1")
-        buf.write("\u09e3\u09f2\u09f3\u0a07\u0a0c\u0a11\u0a12\u0a15\u0a2a")
-        buf.write("\u0a2c\u0a32\u0a34\u0a35\u0a37\u0a38\u0a3a\u0a3b\u0a5b")
-        buf.write("\u0a5e\u0a60\u0a76\u0a87\u0a8f\u0a91\u0a93\u0a95\u0aaa")
-        buf.write("\u0aac\u0ab2\u0ab4\u0ab5\u0ab7\u0abb\u0abf\u0ad2\u0ae2")
-        buf.write("\u0ae3\u0b07\u0b0e\u0b11\u0b12\u0b15\u0b2a\u0b2c\u0b32")
-        buf.write("\u0b34\u0b35\u0b37\u0b3b\u0b3f\u0b63\u0b73\u0b85\u0b87")
-        buf.write("\u0b8c\u0b90\u0b92\u0b94\u0b97\u0b9b\u0b9c\u0b9e\u0bac")
-        buf.write("\u0bb0\u0bbb\u0bd2\u0c0e\u0c10\u0c12\u0c14\u0c2a\u0c2c")
-        buf.write("\u0c3b\u0c3f\u0c8e\u0c90\u0c92\u0c94\u0caa\u0cac\u0cb5")
-        buf.write("\u0cb7\u0cbb\u0cbf\u0ce0\u0ce2\u0ce3\u0cf3\u0cf4\u0d07")
-        buf.write("\u0d0e\u0d10\u0d12\u0d14\u0d3c\u0d3f\u0d50\u0d62\u0d63")
-        buf.write("\u0d7c\u0d81\u0d87\u0d98\u0d9c\u0db3\u0db5\u0dbd\u0dbf")
-        buf.write("\u0dc8\u0e03\u0e32\u0e34\u0e35\u0e42\u0e47\u0e83\u0e84")
-        buf.write("\u0e86\u0e8c\u0e8f\u0e99\u0e9b\u0ea1\u0ea3\u0ea5\u0ea7")
-        buf.write("\u0ea9\u0eac\u0ead\u0eaf\u0eb2\u0eb4\u0eb5\u0ebf\u0ec6")
-        buf.write("\u0ede\u0ee1\u0f02\u0f49\u0f4b\u0f6e\u0f8a\u0f8e\u1002")
-        buf.write("\u102c\u1041\u1057\u105c\u105f\u1063\u1072\u1077\u1083")
-        buf.write("\u1090\u10fc\u10ff\u124a\u124c\u124f\u1252\u1258\u125a")
-        buf.write("\u125f\u1262\u128a\u128c\u128f\u1292\u12b2\u12b4\u12b7")
-        buf.write("\u12ba\u12c0\u12c2\u12c7\u12ca\u12d8\u12da\u1312\u1314")
-        buf.write("\u1317\u131a\u135c\u1382\u1391\u13a2\u13f6\u1403\u166e")
-        buf.write("\u1671\u1681\u1683\u169c\u16a2\u16ec\u16f3\u16fa\u1702")
-        buf.write("\u170e\u1710\u1713\u1722\u1733\u1742\u1753\u1762\u176e")
-        buf.write("\u1770\u1772\u1782\u17b5\u17de\u1844\u1846\u1879\u1882")
-        buf.write("\u18aa\u18ac\u18f7\u1902\u1920\u1952\u196f\u1972\u1976")
-        buf.write("\u1982\u19ad\u19c3\u19c9\u1a02\u1a18\u1a22\u1a56\u1b07")
-        buf.write("\u1b35\u1b47\u1b4d\u1b85\u1ba2\u1bb0\u1bb1\u1bbc\u1be7")
-        buf.write("\u1c02\u1c25\u1c4f\u1c51\u1c5c\u1c79\u1ceb\u1cee\u1cf0")
-        buf.write("\u1cf3\u1cf7\u1cf8\u2137\u213a\u2d32\u2d69\u2d82\u2d98")
-        buf.write("\u2da2\u2da8\u2daa\u2db0\u2db2\u2db8\u2dba\u2dc0\u2dc2")
-        buf.write("\u2dc8\u2dca\u2dd0\u2dd2\u2dd8\u2dda\u2de0\u3008\u303e")
-        buf.write("\u3043\u3098\u30a1\u30fc\u3101\u312f\u3133\u3190\u31a2")
-        buf.write("\u31bc\u31f2\u3201\u3402\u4db7\u4e02\u9fce\ua002\ua016")
-        buf.write("\ua018\ua48e\ua4d2\ua4f9\ua502\ua60d\ua612\ua621\ua62c")
-        buf.write("\ua62d\ua670\ua6e7\ua7f9\ua803\ua805\ua807\ua809\ua80c")
-        buf.write("\ua80e\ua824\ua842\ua875\ua884\ua8b5\ua8f4\ua8f9\ua8fd")
-        buf.write("\ua927\ua932\ua948\ua962\ua97e\ua986\ua9b4\ua9e2\ua9e6")
-        buf.write("\ua9e9\ua9f1\ua9fc\uaa00\uaa02\uaa2a\uaa42\uaa44\uaa46")
-        buf.write("\uaa4d\uaa62\uaa71\uaa73\uaa78\uaa7c\uaab1\uaab3\uaabf")
-        buf.write("\uaac2\uaac4\uaadd\uaade\uaae2\uaaec\uaaf4\uab08\uab0b")
-        buf.write("\uab10\uab13\uab18\uab22\uab28\uab2a\uab30\uabc2\uabe4")
-        buf.write("\uac02\ud7a5\ud7b2\ud7c8\ud7cd\ud7fd\uf902\ufa6f\ufa72")
-        buf.write("\ufadb\ufb1f\ufb2a\ufb2c\ufb38\ufb3a\ufb3e\ufb40\ufbb3")
-        buf.write("\ufbd5\ufd3f\ufd52\ufd91\ufd94\ufdc9\ufdf2\ufdfd\ufe72")
-        buf.write("\ufe76\ufe78\ufefe\uff68\uff71\uff73\uff9f\uffa2\uffc0")
-        buf.write("\uffc4\uffc9\uffcc\uffd1\uffd4\uffd9\uffdc\uffde'\2\62")
-        buf.write(";\u0662\u066b\u06f2\u06fb\u07c2\u07cb\u0968\u0971\u09e8")
-        buf.write("\u09f1\u0a68\u0a71\u0ae8\u0af1\u0b68\u0b71\u0be8\u0bf1")
-        buf.write("\u0c68\u0c71\u0ce8\u0cf1\u0d68\u0d71\u0de8\u0df1\u0e52")
-        buf.write("\u0e5b\u0ed2\u0edb\u0f22\u0f2b\u1042\u104b\u1092\u109b")
-        buf.write("\u17e2\u17eb\u1812\u181b\u1948\u1951\u19d2\u19db\u1a82")
-        buf.write("\u1a8b\u1a92\u1a9b\u1b52\u1b5b\u1bb2\u1bbb\u1c42\u1c4b")
-        buf.write("\u1c52\u1c5b\ua622\ua62b\ua8d2\ua8db\ua902\ua90b\ua9d2")
-        buf.write("\ua9db\ua9f2\ua9fb\uaa52\uaa5b\uabf2\uabfb\uff12\uff1b")
-        buf.write("\2{\2\7\3\2\2\2\2\t\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2\2")
-        buf.write("\17\3\2\2\2\2\21\3\2\2\2\2\23\3\2\2\2\2\25\3\2\2\2\2\27")
-        buf.write("\3\2\2\2\3'\3\2\2\2\5+\3\2\2\2\7.\3\2\2\2\t\60\3\2\2")
-        buf.write("\2\13\62\3\2\2\2\r\66\3\2\2\2\17D\3\2\2\2\21I\3\2\2\2")
-        buf.write("\23M\3\2\2\2\25Y\3\2\2\2\27_\3\2\2\2\31f\3\2\2\2\33h\3")
-        buf.write("\2\2\2\35j\3\2\2\2\37l\3\2\2\2!n\3\2\2\2#p\3\2\2\2%r\3")
-        buf.write("\2\2\2'(\t\2\2\2()\t\3\2\2)*\t\4\2\2*\4\3\2\2\2+,\t\5")
-        buf.write("\2\2,-\t\4\2\2-\6\3\2\2\2./\7*\2\2/\b\3\2\2\2\60\61\7")
-        buf.write("+\2\2\61\n\3\2\2\2\62\63\t\3\2\2\63\64\t\6\2\2\64\65\t")
-        buf.write("\7\2\2\65\f\3\2\2\2\66\67\t\b\2\2\678\t\t\2\28\16\3\2")
-        buf.write("\2\29E\7?\2\2:;\7#\2\2;E\7?\2\2<E\7\u0080\2\2=>\7@\2\2")
-        buf.write(">E\7?\2\2?@\7>\2\2@E\7?\2\2AE\t\n\2\2BE\5\3\2\2CE\5\5")
-        buf.write("\3\2D9\3\2\2\2D:\3\2\2\2D<\3\2\2\2D=\3\2\2\2D?\3\2\2\2")
-        buf.write("DA\3\2\2\2DB\3\2\2\2DC\3\2\2\2E\20\3\2\2\2FJ\5\31\r\2")
-        buf.write("GJ\5%\23\2HJ\t\13\2\2IF\3\2\2\2IG\3\2\2\2IH\3\2\2\2JK")
-        buf.write("\3\2\2\2KI\3\2\2\2KL\3\2\2\2L\22\3\2\2\2MS\7$\2\2NR\n")
-        buf.write("\f\2\2OP\7^\2\2PR\7$\2\2QN\3\2\2\2QO\3\2\2\2RU\3\2\2\2")
-        buf.write("SQ\3\2\2\2ST\3\2\2\2TV\3\2\2\2US\3\2\2\2VW\7$\2\2W\24")
-        buf.write("\3\2\2\2XZ\t\r\2\2YX\3\2\2\2Z[\3\2\2\2[Y\3\2\2\2[\\\3")
-        buf.write("\2\2\2\\]\3\2\2\2]^\b\13\2\2^\26\3\2\2\2_`\13\2\2\2`\30")
-        buf.write("\3\2\2\2ag\5\33\16\2bg\5\35\17\2cg\5\37\20\2dg\5!\21\2")
-        buf.write("eg\5#\22\2fa\3\2\2\2fb\3\2\2\2fc\3\2\2\2fd\3\2\2\2fe\3")
-        buf.write("\2\2\2g\32\3\2\2\2hi\t\16\2\2i\34\3\2\2\2jk\t\17\2\2k")
-        buf.write('\36\3\2\2\2lm\t\20\2\2m \3\2\2\2no\t\21\2\2o"\3\2\2\2')
-        buf.write("pq\t\22\2\2q$\3\2\2\2rs\t\23\2\2s&\3\2\2\2\n\2DIKQS[f")
-        buf.write("\3\b\2\2")
-        return buf.getvalue()
-
-
-class ContactQLLexer(Lexer):
-
-    atn = ATNDeserializer().deserialize(serializedATN())
-
-    decisionsToDFA = [DFA(ds, i) for i, ds in enumerate(atn.decisionToState)]
-
-    LPAREN = 1
-    RPAREN = 2
-    AND = 3
-    OR = 4
-    COMPARATOR = 5
-    TEXT = 6
-    STRING = 7
-    WS = 8
-    ERROR = 9
-
-    channelNames = ["DEFAULT_TOKEN_CHANNEL", "HIDDEN"]
-
-    modeNames = ["DEFAULT_MODE"]
-
-    literalNames = ["<INVALID>", "'('", "')'"]
-
-    symbolicNames = ["<INVALID>", "LPAREN", "RPAREN", "AND", "OR", "COMPARATOR", "TEXT", "STRING", "WS", "ERROR"]
-
-    ruleNames = [
-        "HAS",
-        "IS",
-        "LPAREN",
-        "RPAREN",
-        "AND",
-        "OR",
-        "COMPARATOR",
-        "TEXT",
-        "STRING",
-        "WS",
-        "ERROR",
-        "UnicodeLetter",
-        "UnicodeClass_LU",
-        "UnicodeClass_LL",
-        "UnicodeClass_LT",
-        "UnicodeClass_LM",
-        "UnicodeClass_LO",
-        "UnicodeDigit",
-    ]
-
-    grammarFileName = "ContactQL.g4"
-
-    def __init__(self, input=None, output: TextIO = sys.stdout):
-        super().__init__(input, output)
-        self.checkVersion("4.7.2")
-        self._interp = LexerATNSimulator(self, self.atn, self.decisionsToDFA, PredictionContextCache())
-        self._actions = None
-        self._predicates = None
