--- temba/contacts/search/gen/ContactQLLexer.tokens
@@ -1,11 +1,0 @@
-LPAREN=1
-RPAREN=2
-AND=3
-OR=4
-COMPARATOR=5
-TEXT=6
-STRING=7
-WS=8
-ERROR=9
-'('=1
-')'=2
