+++ temba/contacts/tests.py
@@ -15,7 +15,6 @@ from django.core.validators import ValidationError
 from django.db import connection
 from django.db.models import Value as DbValue
 from django.db.models.functions import Concat, Substr
-from django.test import TestCase
 from django.test.utils import override_settings
 from django.urls import reverse
 from django.utils import timezone
@@ -25,7 +24,7 @@ from temba.api.models import WebHookResult
 from temba.campaigns.models import Campaign, CampaignEvent, EventFire
 from temba.channels.models import Channel, ChannelEvent, ChannelLog
 from temba.contacts.models import DELETED_SCHEME
-from temba.contacts.search import SearchException, SearchResults, evaluate_query, is_phonenumber, search_contacts
+from temba.contacts.search import SearchException, SearchResults, search_contacts
 from temba.contacts.views import ContactListView
 from temba.flows.models import Flow, FlowRun
 from temba.ivr.models import IVRCall
@@ -61,7 +60,6 @@ from .models import (
     ContactURN,
     ExportContactsTask,
 )
-from .search import BoolCombination, Condition, ContactQuery, IsSetCondition, SinglePropCombination, legacy_parse_query
 from .tasks import check_elasticsearch_lag, squash_contactgroupcounts
 from .templatetags.contacts import contact_field, history_class, history_icon
 
@@ -204,7 +202,7 @@ class ContactCRUDLTest(TembaTest):
         self.assertIn(joe, survey_audience.contacts.all())
 
         self.assertEqual(
-            call(self.org.id, str(active_contacts.uuid), "Joe", "", offset=0, exclude_ids=[]),
+            call(self.org.id, group_uuid=str(active_contacts.uuid), query="Joe", sort="", offset=0, exclude_ids=[]),
             mr_mocks.calls["contact_search"][-1],
         )
 
@@ -213,7 +211,9 @@ class ContactCRUDLTest(TembaTest):
 
         # we re-run the search for the response, but exclude Joe
         self.assertEqual(
-            call(self.org.id, str(active_contacts.uuid), "Joe", "", offset=0, exclude_ids=[joe.id]),
+            call(
+                self.org.id, group_uuid=str(active_contacts.uuid), query="Joe", sort="", offset=0, exclude_ids=[joe.id]
+            ),
             mr_mocks.calls["contact_search"][-1],
         )
 
@@ -1652,683 +1652,6 @@ class ContactTest(TembaTest):
             self.assertEqual(self.joe.get_field_value(nick), "Joey")
             self.assertIsNone(self.frank.get_field_value(nick))
             self.assertIsNone(self.billy.get_field_value(nick))
-
-    def test_contact_search_evaluation_created_on_utc_rollover(self):
-        # org is in Africa/Kigali timezone: +02:00
-        self.joe.created_on = datetime(2019, 6, 8, 23, 14, 0, tzinfo=pytz.UTC)
-        self.joe.save(update_fields=("created_on",))
-
-        query_created_on = self.joe.created_on.astimezone(self.org.timezone).date().isoformat()
-
-        # date in org timezone is the 'next' day
-        self.assertEqual(query_created_on, "2019-06-09")
-
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on = "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_created_on = (self.joe.created_on - timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on > "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on >= "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_created_on = (self.joe.created_on + timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on < "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on <= "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-
-    @mock_mailroom
-    def test_contact_search_evaluation(self, mr_mocks):
-        self.setUpLocations()
-
-        ContactField.get_or_create(self.org, self.admin, "gender", "Gender", value_type=Value.TYPE_TEXT)
-        ContactField.get_or_create(self.org, self.admin, "age", "Age", value_type=Value.TYPE_NUMBER)
-        ContactField.get_or_create(self.org, self.admin, "joined", "Joined On", value_type=Value.TYPE_DATETIME)
-        ContactField.get_or_create(self.org, self.admin, "ward", "Ward", value_type=Value.TYPE_WARD)
-        ContactField.get_or_create(self.org, self.admin, "district", "District", value_type=Value.TYPE_DISTRICT)
-        ContactField.get_or_create(self.org, self.admin, "state", "State", value_type=Value.TYPE_STATE)
-
-        # test 'uuid' attribute
-        self.assertTrue(evaluate_query(self.org, f'uuid = "{self.joe.uuid}"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(
-            evaluate_query(self.org, f'uuid = "{self.frank.uuid}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertFalse(
-            evaluate_query(self.org, f'uuid != "{self.joe.uuid}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(evaluate_query(self.org, 'uuid != "123456"', contact_json=self.joe.as_search_json()))
-
-        # uuid does not support `has` operator or set checks
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'uuid ~ "123"', contact_json=self.joe.as_search_json()
-        )
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'uuid = ""', contact_json=self.joe.as_search_json()
-        )
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'uuid != ""', contact_json=self.joe.as_search_json()
-        )
-
-        # test 'name' attribute
-        self.assertTrue(evaluate_query(self.org, 'name = "Joe Blow"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "name = Joe", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "name ~ blow", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'name = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'name != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'name != "Bob"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'name != "Joe Blow"', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'name = ""', contact_json={}))
-        self.assertFalse(evaluate_query(self.org, 'name != ""', contact_json={}))
-        # nothing to compare
-        self.assertFalse(evaluate_query(self.org, "name = Joe", contact_json={}))
-
-        # 'voldemort' does not have any attributes, evaluates to True
-        self.assertTrue(evaluate_query(self.org, 'name != "Joe Blow"', contact_json=self.voldemort.as_search_json()))
-
-        # test 'language' attribute
-        self.joe.language = "eng"
-        self.joe.save(update_fields=("language",))
-        self.assertTrue(evaluate_query(self.org, 'language = "eng"', contact_json=self.joe.as_search_json()))
-
-        self.assertFalse(evaluate_query(self.org, 'language != "eng"', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'language != "ita"', contact_json=self.joe.as_search_json()))
-
-        self.assertFalse(evaluate_query(self.org, 'language = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'language != ""', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'language != "bla"', contact_json=self.voldemort.as_search_json()))
-
-        # language does not support `has` operator
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'language ~ "eng"', contact_json=self.joe.as_search_json()
-        )
-
-        self.joe.language = None
-        self.joe.save(update_fields=("language",))
-
-        self.assertFalse(evaluate_query(self.org, 'language = "eng"', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'language = ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'language != ""', contact_json=self.joe.as_search_json()))
-
-        # test 'created_on' attribute
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'created_on = "this-is-not-a-date"',
-            contact_json=self.joe.as_search_json(),
-        )
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'created_on ~ "2016-01-01"',
-            contact_json=self.joe.as_search_json(),
-        )
-        query_created_on = self.joe.created_on.astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on = "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_created_on = (self.joe.created_on - timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on > "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on >= "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_created_on = (self.joe.created_on + timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on < "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'created_on <= "{query_created_on}"', contact_json=self.joe.as_search_json())
-        )
-
-        # test 'last_seen_on' attribute
-        self.joe.last_seen_on = datetime(2020, 3, 17, 13, 0, 0, 0, pytz.UTC)
-        self.joe.save(update_fields=("last_seen_on",))
-
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'last_seen_on = "this-is-not-a-date"',
-            contact_json=self.joe.as_search_json(),
-        )
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'last_seen_on ~ "2016-01-01"',
-            contact_json=self.joe.as_search_json(),
-        )
-        query_last_seen_on = self.joe.last_seen_on.astimezone(self.org.timezone).date().isoformat()
-        self.assertTrue(
-            evaluate_query(self.org, f'last_seen_on = "{query_last_seen_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_last_seen_on = (
-            (self.joe.last_seen_on - timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'last_seen_on > "{query_last_seen_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'last_seen_on >= "{query_last_seen_on}"', contact_json=self.joe.as_search_json())
-        )
-        query_last_seen_on = (
-            (self.joe.last_seen_on + timedelta(days=6)).astimezone(self.org.timezone).date().isoformat()
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'last_seen_on < "{query_last_seen_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(self.org, f'last_seen_on <= "{query_last_seen_on}"', contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(evaluate_query(self.org, f'last_seen_on != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, f'last_seen_on = ""', contact_json=self.joe.as_search_json()))
-
-        self.joe.last_seen_on = None
-        self.joe.save(update_fields=("last_seen_on",))
-
-        self.assertFalse(
-            evaluate_query(self.org, f"last_seen_on = 2016-01-01", contact_json=self.joe.as_search_json())
-        )
-
-        # test TEXT field type
-        self.assertFalse(evaluate_query(self.org, 'gender != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'gender = ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "gender = male", contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "gender", "Male")
-        self.assertTrue(evaluate_query(self.org, "gender = male", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "gender = Female", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'gender != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'gender = ""', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, "gender != female", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "gender != male", contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, "gender != female", contact_json=self.voldemort.as_search_json()))
-
-        # test DECIMAL field type
-        self.assertFalse(evaluate_query(self.org, 'age != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'age = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "age", "cedevita is not a number")
-        self.assertFalse(evaluate_query(self.org, "age < 99", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'age != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'age = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "age", "18")
-        self.assertTrue(evaluate_query(self.org, 'age != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'age = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "age = 18", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "age > 20", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "age >= 15", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "age <= 17", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "age < 20", contact_json=self.joe.as_search_json()))
-
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'age < "cedevita is not a number"',
-            contact_json=self.joe.as_search_json(),
-        )
-
-        # test DATETIME field type
-        self.assertFalse(evaluate_query(self.org, 'joined != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'joined = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "joined", "cedevita is not a datetime object")
-        self.assertFalse(evaluate_query(self.org, "joined < 01-04-2018", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'joined != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'joined = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "joined", "01-03-2018")
-        self.assertTrue(evaluate_query(self.org, 'joined != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'joined = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "joined = 01-03-2018", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "joined > 01-04-2018", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "joined >= 01-03-2018", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "joined <= 28-02-2018", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "joined < 01-04-2018", contact_json=self.joe.as_search_json()))
-        self.assertRaises(
-            SearchException,
-            evaluate_query,
-            self.org,
-            'joined < "cedevita is not a datetime object"',
-            contact_json=self.joe.as_search_json(),
-        )
-
-        # test WARD field type
-        self.assertFalse(evaluate_query(self.org, 'ward != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'ward = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "ward", "cedevita is not a ward")
-        self.assertFalse(evaluate_query(self.org, 'ward != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'ward = ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'ward = "cedevita"', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "ward", "Rwanda > Eastern Province > Rwamagana > Bukure")
-        self.assertTrue(evaluate_query(self.org, 'ward != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'ward = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'ward = "bUKuRE"', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'ward != "Rwamagana"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'ward != "Bukure"', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'ward != "Rwamagana"', contact_json=self.voldemort.as_search_json()))
-
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'ward ~ "ukur"', contact_json=self.joe.as_search_json()
-        )
-
-        self.assertFalse(
-            evaluate_query(self.org, 'ward = "cedevita is not a ward"', contact_json=self.joe.as_search_json())
-        )
-
-        # test DISTRICT field type
-        self.assertFalse(evaluate_query(self.org, 'district != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'district = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "district", "cedevita is not a district")
-        self.assertFalse(evaluate_query(self.org, 'district != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'district = ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'district = "cedevita"', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "district", "Rwanda > Eastern Province > Rwamagana")
-        self.assertTrue(evaluate_query(self.org, 'district != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'district = ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'district = "Rwamagana"', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'district != "Bukure"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'district != "Rwamagana"', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(
-            evaluate_query(self.org, 'district != "Rwamagana"', contact_json=self.voldemort.as_search_json())
-        )
-
-        self.assertFalse(
-            evaluate_query(self.org, 'district = "cedevita is not a district"', contact_json=self.joe.as_search_json())
-        )
-
-        # test STATE field type
-        self.assertFalse(evaluate_query(self.org, 'state != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'state = ""', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "state", "cedevita is not a state")
-        self.assertFalse(evaluate_query(self.org, 'state != ""', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'state = ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'state = "cedevita"', contact_json=self.joe.as_search_json()))
-
-        self.set_contact_field(self.joe, "state", "Rwanda > Eastern Province")
-        self.assertTrue(evaluate_query(self.org, 'state != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'state = ""', contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(
-            evaluate_query(self.org, 'state != "Western Province"', contact_json=self.joe.as_search_json())
-        )
-        self.assertFalse(
-            evaluate_query(self.org, 'state != "Eastern Province"', contact_json=self.joe.as_search_json())
-        )
-
-        self.assertTrue(
-            evaluate_query(self.org, 'state != "Western Province"', contact_json=self.voldemort.as_search_json())
-        )
-
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'state ~ "stern"', contact_json=self.joe.as_search_json()
-        )
-
-        self.assertFalse(
-            evaluate_query(self.org, 'state = "cedevita is not a state"', contact_json=self.joe.as_search_json())
-        )
-
-        # URN is set
-        self.assertTrue(evaluate_query(self.org, 'tel != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'twitter = ""', contact_json=self.joe.as_search_json()))
-
-        self.assertFalse(evaluate_query(self.org, 'tel != ""', contact_json=self.billy.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'twitter = ""', contact_json=self.billy.as_search_json()))
-
-        # add another tel URN
-        mods = self.joe.update_urns(["tel:+250781111999", "tel:+250781111111", "twitter:blow80"])
-        self.joe.modify(self.user, mods)
-
-        self.assertTrue(evaluate_query(self.org, "+250781111111", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "tel = +250781111999", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "tel has 278", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'twitter = "blow80"', contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, 'twitter has "blow"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'twitter has "joe"', contact_json=self.joe.as_search_json()))
-
-        # test 'urn' attribute
-        self.assertTrue(evaluate_query(self.org, "urn = +250781111111", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "urn = +250781111999", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "urn = blow80", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "urn = +250781111222", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "urn != +250781111111", contact_json=self.joe.as_search_json()))
-        self.assertTrue(evaluate_query(self.org, "urn != +250781111222", contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, "urn ~ 999", contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, "urn ~ 8888", contact_json=self.joe.as_search_json()))
-
-        self.assertTrue(evaluate_query(self.org, 'urn != ""', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, 'urn = ""', contact_json=self.joe.as_search_json()))
-
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'urn > "x"', contact_json=self.joe.as_search_json()
-        )
-
-        self.assertTrue(
-            evaluate_query(self.org, "joined = 01-03-2018 AND age < 19", contact_json=self.joe.as_search_json())
-        )
-        self.assertFalse(
-            evaluate_query(self.org, "joined = 01-03-2018 AND age > 19", contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(
-                self.org, "joined = 01-03-2018 AND tel = +250781111111", contact_json=self.joe.as_search_json()
-            )
-        )
-        self.assertTrue(
-            evaluate_query(self.org, "joined = 01-03-2018 OR age > 19", contact_json=self.joe.as_search_json())
-        )
-        self.assertTrue(
-            evaluate_query(
-                self.org,
-                'joined = 01-03-2018 AND (age > 19 OR gender = "male")',
-                contact_json=self.joe.as_search_json(),
-            )
-        )
-
-        # values can contain quotes
-        self.set_contact_field(self.joe, "gender", 'M"F')
-        self.assertTrue(evaluate_query(self.org, r'gender = "M\"F"', contact_json=self.joe.as_search_json()))
-        self.assertFalse(evaluate_query(self.org, r'gender != "M\"F"', contact_json=self.joe.as_search_json()))
-        self.set_contact_field(self.joe, "gender", "male")
-
-        # non-existent field or attribute
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, "credits > 10", contact_json=self.joe.as_search_json()
-        )
-
-        # unsupported text operator
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'gender ~ "cedevita"', contact_json=self.joe.as_search_json()
-        )
-
-        # unsupported number operator
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, "age ~ 13", contact_json=self.joe.as_search_json()
-        )
-
-        # unsupported date operator
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, "joined ~ 01-03-2018", contact_json=self.joe.as_search_json()
-        )
-
-        # unsupported scheme operator
-        self.assertRaises(
-            SearchException, evaluate_query, self.org, 'tel > "cedevita"', contact_json=self.joe.as_search_json()
-        )
-
-        with AnonymousOrg(self.org):
-            self.joe.refresh_from_db()
-
-            self.assertTrue(evaluate_query(self.org, "gender = male", contact_json=self.joe.as_search_json()))
-            self.assertTrue(evaluate_query(self.org, "gender != female", contact_json=self.joe.as_search_json()))
-            self.assertTrue(evaluate_query(self.org, "age >= 15", contact_json=self.joe.as_search_json()))
-
-            # do not evaluate URN queries if org is anonymous
-            self.assertRaises(
-                SearchException,
-                evaluate_query,
-                self.org,
-                "tel = +250781111111",
-                contact_json=self.joe.as_search_json(),
-            )
-            self.assertRaises(
-                SearchException,
-                evaluate_query,
-                self.org,
-                "urn = +250781111111",
-                contact_json=self.joe.as_search_json(),
-            )
-            self.assertRaises(
-                SearchException,
-                evaluate_query,
-                self.org,
-                "joined = 01-03-2018 AND tel = +250781111111",
-                contact_json=self.joe.as_search_json(),
-            )
-
-            # URN existence checks allowed
-            self.assertFalse(evaluate_query(self.org, 'tel = ""', contact_json=self.joe.as_search_json()))
-            self.assertTrue(evaluate_query(self.org, 'tel != ""', contact_json=self.joe.as_search_json()))
-            self.assertFalse(evaluate_query(self.org, 'urn = ""', contact_json=self.joe.as_search_json()))
-            self.assertTrue(evaluate_query(self.org, 'urn != ""', contact_json=self.joe.as_search_json()))
-
-            # this will be parsed as search for contact id
-            self.assertRaises(
-                SearchException, evaluate_query, self.org, str(self.joe.pk), contact_json=self.joe.as_search_json()
-            )
-
-    def test_contact_search_parsing(self):
-        # implicit condition on name
-        self.assertEqual(legacy_parse_query("will"), ContactQuery(Condition("name", "~", "will")))
-        self.assertEqual(legacy_parse_query("1will2"), ContactQuery(Condition("name", "~", "1will2")))
-
-        self.assertEqual(legacy_parse_query("will").as_text(), 'name ~ "will"')
-        self.assertEqual(legacy_parse_query("1will2").as_text(), 'name ~ "1will2"')
-
-        # implicit condition on tel if value is all tel chars
-        self.assertEqual(legacy_parse_query("1234"), ContactQuery(Condition("tel", "~", "1234")))
-        self.assertEqual(legacy_parse_query("+12-34"), ContactQuery(Condition("tel", "~", "1234")))
-        self.assertEqual(legacy_parse_query("1234", as_anon=True), ContactQuery(Condition("id", "=", "1234")))
-        self.assertEqual(legacy_parse_query("+12-34", as_anon=True), ContactQuery(Condition("name", "~", "+12-34")))
-        self.assertEqual(legacy_parse_query("bob", as_anon=True), ContactQuery(Condition("name", "~", "bob")))
-
-        self.assertEqual(legacy_parse_query("1234").as_text(), "tel ~ 1234")
-        self.assertEqual(legacy_parse_query("+12-34").as_text(), "tel ~ 1234")
-
-        # boolean combinations of implicit conditions
-        self.assertEqual(
-            legacy_parse_query("will felix", optimize=False),
-            ContactQuery(
-                BoolCombination(BoolCombination.AND, Condition("name", "~", "will"), Condition("name", "~", "felix"))
-            ),
-        )
-        self.assertEqual(
-            legacy_parse_query("will felix"),
-            ContactQuery(
-                SinglePropCombination(
-                    "name", BoolCombination.AND, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                )
-            ),
-        )
-        self.assertEqual(
-            legacy_parse_query("will and felix", optimize=False),
-            ContactQuery(
-                BoolCombination(BoolCombination.AND, Condition("name", "~", "will"), Condition("name", "~", "felix"))
-            ),
-        )
-        self.assertEqual(
-            legacy_parse_query("will or felix or matt", optimize=False),
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.OR,
-                    BoolCombination(
-                        BoolCombination.OR, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                    ),
-                    Condition("name", "~", "matt"),
-                )
-            ),
-        )
-
-        # property conditions
-        self.assertEqual(legacy_parse_query("name=will"), ContactQuery(Condition("name", "=", "will")))
-        self.assertEqual(legacy_parse_query('name ~ "felix"'), ContactQuery(Condition("name", "~", "felix")))
-        self.assertEqual(legacy_parse_query('name != "felix"'), ContactQuery(Condition("name", "!=", "felix")))
-
-        # empty string conditions
-        self.assertEqual(legacy_parse_query('name is ""'), ContactQuery(IsSetCondition("name", "is")))
-        self.assertEqual(legacy_parse_query('name!=""'), ContactQuery(IsSetCondition("name", "!=")))
-
-        # boolean combinations of property conditions
-        self.assertEqual(
-            legacy_parse_query('name=will or name ~ "felix"', optimize=False),
-            ContactQuery(
-                BoolCombination(BoolCombination.OR, Condition("name", "=", "will"), Condition("name", "~", "felix"))
-            ),
-        )
-        self.assertEqual(
-            legacy_parse_query('name=will or name ~ "felix"'),
-            ContactQuery(
-                SinglePropCombination(
-                    "name", BoolCombination.OR, Condition("name", "=", "will"), Condition("name", "~", "felix")
-                )
-            ),
-        )
-
-        # mixture of simple and property conditions
-        self.assertEqual(
-            legacy_parse_query('will or name ~ "felix"'),
-            ContactQuery(
-                SinglePropCombination(
-                    "name", BoolCombination.OR, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                )
-            ),
-        )
-
-        # optimization will merge conditions combined with the same op
-        self.assertEqual(
-            legacy_parse_query("will or felix or matt"),
-            ContactQuery(
-                SinglePropCombination(
-                    "name",
-                    BoolCombination.OR,
-                    Condition("name", "~", "will"),
-                    Condition("name", "~", "felix"),
-                    Condition("name", "~", "matt"),
-                )
-            ),
-        )
-
-        # but not conditions combined with different ops
-        self.assertEqual(
-            legacy_parse_query("will or felix and matt"),
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.OR,
-                    Condition("name", "~", "will"),
-                    SinglePropCombination(
-                        "name", BoolCombination.AND, Condition("name", "~", "felix"), Condition("name", "~", "matt")
-                    ),
-                )
-            ),
-        )
-
-        # optimization respects explicit precedence defined with parentheses
-        self.assertEqual(
-            legacy_parse_query("(will or felix) and matt"),
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.AND,
-                    SinglePropCombination(
-                        "name", BoolCombination.OR, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                    ),
-                    Condition("name", "~", "matt"),
-                )
-            ),
-        )
-
-        # implicit ANDing of conditions
-        query = legacy_parse_query('will felix name ~ "matt"')
-        self.assertEqual(
-            query,
-            ContactQuery(
-                SinglePropCombination(
-                    "name",
-                    BoolCombination.AND,
-                    Condition("name", "~", "will"),
-                    Condition("name", "~", "felix"),
-                    Condition("name", "~", "matt"),
-                )
-            ),
-        )
-        self.assertEqual(query.as_text(), 'name ~ "will" AND name ~ "felix" AND name ~ "matt"')
-
-        self.assertEqual(
-            legacy_parse_query('will felix name ~ "matt"', optimize=False),
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.AND,
-                    BoolCombination(
-                        BoolCombination.AND, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                    ),
-                    Condition("name", "~", "matt"),
-                )
-            ),
-        )
-
-        # boolean operator precedence is AND before OR, even when AND is implicit
-        self.assertEqual(
-            legacy_parse_query("will and felix or matt amber", optimize=False),
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.OR,
-                    BoolCombination(
-                        BoolCombination.AND, Condition("name", "~", "will"), Condition("name", "~", "felix")
-                    ),
-                    BoolCombination(
-                        BoolCombination.AND, Condition("name", "~", "matt"), Condition("name", "~", "amber")
-                    ),
-                )
-            ),
-        )
-
-        # boolean combinations can themselves be combined
-        query = legacy_parse_query('(Age < 18 and Gender = "male") or (Age > 18 and Gender = "female")')
-        self.assertEqual(
-            query,
-            ContactQuery(
-                BoolCombination(
-                    BoolCombination.OR,
-                    BoolCombination(
-                        BoolCombination.AND, Condition("age", "<", "18"), Condition("gender", "=", "male")
-                    ),
-                    BoolCombination(
-                        BoolCombination.AND, Condition("age", ">", "18"), Condition("gender", "=", "female")
-                    ),
-                )
-            ),
-        )
-        self.assertEqual(query.as_text(), '(age < 18 AND gender = "male") OR (age > 18 AND gender = "female")')
-
-        self.assertEqual(str(legacy_parse_query('Age < 18 and Gender = "male"')), "AND(age<18, gender=male)")
-        self.assertEqual(str(legacy_parse_query("Age > 18 and Age < 30")), "AND[age](>18, <30)")
-
-        # query with UTF-8 characters (non-ascii)
-        query = legacy_parse_query('district="Kayônza"')
-        self.assertEqual(query.as_text(), 'district = "Kayônza"')
-
-        # query that has @ sign
-        query = legacy_parse_query('email ~ "user@example.com"')
-        self.assertEqual(query.as_text(), 'email ~ "user@example.com"')
-
-        query = legacy_parse_query("email ~ user@example.com")
-        self.assertEqual(query.as_text(), 'email ~ "user@example.com"')
-
-        # escaped quotes
-        query = legacy_parse_query(r'name ~ "O\"Learly"')
-        self.assertEqual(query.as_text(), r'name ~ "O"Learly"')
 
     @mock_mailroom
     def test_omnibox(self, mr_mocks):
@@ -5707,24 +5030,6 @@ class URNTest(TembaTest):
         self.assertTrue(URN.validate("vk:12345678901234567"))
 
 
-class PhoneNumberTest(TestCase):
-    def test_is_phonenumber(self):
-        # these should match as phone numbers
-        self.assertEqual(is_phonenumber("+12345678901"), (True, "12345678901"))
-        self.assertEqual(is_phonenumber("+1-234-567-8901"), (True, "12345678901"))
-        self.assertEqual(is_phonenumber("+1 (234) 567-8901"), (True, "12345678901"))
-        self.assertEqual(is_phonenumber("+12345678901"), (True, "12345678901"))
-        self.assertEqual(is_phonenumber("+12 34 567 8901"), (True, "12345678901"))
-        self.assertEqual(is_phonenumber(" 234 567 8901 "), (True, "2345678901"))
-
-        # these should not be parsed as numbers
-        self.assertEqual(is_phonenumber("+12345678901 not a number"), (False, None))
-        self.assertEqual(is_phonenumber(""), (False, None))
-        self.assertEqual(is_phonenumber("AMAZONS"), (False, None))
-        self.assertEqual(is_phonenumber('name = "Jack"'), (False, None))
-        self.assertEqual(is_phonenumber('(social = "234-432-324")'), (False, None))
-
-
 class ESIntegrationTest(TembaNonAtomicTest):
     def test_ES_contacts_index(self):
         self.create_anonymous_user()
@@ -5804,7 +5109,7 @@ class ESIntegrationTest(TembaNonAtomicTest):
             self.create_contact(name, urns=urns, fields=fields)
 
         def q(query):
-            results = search_contacts(self.org.id, self.org.cached_active_contacts_group.uuid, query, None)
+            results = search_contacts(self.org, query, group=self.org.cached_active_contacts_group)
             return results.total
 
         db_config = connection.settings_dict
